Rename api module to mocks
